requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pyzmq>=26.0.0
pydantic>=2.7.0
structlog==24.1.0
pyyaml==6.0.1
python-dotenv==1.0.1
openai==1.30.1
pytest==8.2.1
ruff==0.4.7
